Move helper traits classes out of Quantity.hh
